add MyPastesPage

+++ backend/src/services/paste.service.js
@@ -2,13 +2,16 @@ const PasteRepository = require("../repositories/paste.repository");
 const AccountRepository = require("../repositories/account.repository");
 const { generateRandomString } = require("../utils/randomGenerator");
 const ValidationError = require("../errors/ValidationError");
+const ForbiddenError = require("../errors/ForbiddenError");
+const NotFoundError = require("../errors/NotFoundError");
 
 const PasteService = {
   createPaste: async (pasteData) => {
-    let randomId = generateRandomString();
+    let randomId;
 
-    if (await PasteRepository.findById(randomId))
+    do {
       randomId = generateRandomString();
+    } while (await PasteRepository.findById(randomId))
 
     // if the paste has an account number, verify if that account exists
     if (pasteData.accountNumber) {
@@ -25,14 +28,28 @@ const PasteService = {
   },
 
   getPasteById: async (id) => {
-    return await PasteRepository.findById(id);
+    const paste = await PasteRepository.findById(id)
+
+    if (!paste) throw new NotFoundError("Paste not found.")
+
+    return paste;
   },
 
   getPastesByAccount: async (accountNumber) => {
     return await PasteRepository.findByAccountNumber(accountNumber);
   },
 
-  deletePaste: async (pasteId) => {
+  deletePaste: async (pasteId, accountNumber) => {
+    const originalPaste = await PasteRepository.findById(pasteId);
+
+    if (!accountNumber) {
+      throw new ValidationError("You didn't provide an Account Number!")
+    }
+
+    if (originalPaste.accountNumber && originalPaste.accountNumber != accountNumber) {
+      throw new ForbiddenError("You do not have the permission to delete this paste!");
+    }
+
     return await PasteRepository.deleteById(pasteId);
   },
 };
